## #️⃣연관된 이슈

> #29 

## 작업 상세 내용

FeedPage
- 메인 페이지 프리페치 쿼리의 키 불일치 문제 해결
- 프리페치 데이터가 제대로 캐싱되도록 쿼리 키 일치화 적용
- Lighthouse  LCP 1.2s → 0.8s, Speed Index 0.8s → 0.5s로 개선 확인
- 프리페치가 정상 작동하여 화면 깜박임이 줄어듦

PostDetailPage
- prefetchQuery 사용하여 미리 데이터 가져오도록 개선
- Lighthouse 성능 81점→ 98점, LCP  1.0s → 0.8s, Speed Index 0.7s → 0.4s로 개선 확인
- 이미지 최적화 등 추가적인 성능 개선 필요

## 참고할만한 자료(선택)